Revert ""
This reverts commit 06467ea58eb50dfaf4256e3dd527d7f7e6c57b0b.

+++ src/pasteleriadaa/PasteleriaDAA.java
@@ -1,13 +1,6 @@
 
 package pasteleriadaa;
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.util.Comparator;
+import java.io.*;
 import java.util.LinkedList;
 import java.util.PriorityQueue;
 
@@ -22,14 +15,12 @@ public class PasteleriaDAA {
     public static void main(String[] args) throws IOException {
         if (args.length > 0) {
             cargar(args[0]);
-            RyP();
+            RyP ();
             if (args.length == 2) 
                 guardar(args[1]);
             else mostrarPantalla(); 
         } else System.err.println("Introduce fichero de entrada.");
-        
     }
-    
     
     public static void cargar(String fichero) throws FileNotFoundException, IOException{
         String[] aux;
@@ -62,9 +53,9 @@ public class PasteleriaDAA {
         //Comprobamos si el fichero existe
         File archivo = new File(fichero);
         BufferedWriter bw;
-//        if (archivo.exists()) {
-//            System.err.println("El fichero de salida ya existe");
-//        } else {
+        if (archivo.exists()) {
+            System.err.println("El fichero de salida ya existe");
+        } else {
             try{
                 bw = new BufferedWriter(new FileWriter(archivo));
                 for (int i = 0; i < resultado.size() - 1; i++) {
@@ -77,14 +68,14 @@ public class PasteleriaDAA {
             } catch (Exception ex) {
                 System.err.println("Error Salida: " + ex.getMessage());
             }
-//        }
+        }
     }
 
     private static void mostrarPantalla() {
         for (int i = 0; i < resultado.size()-1; i++) {
-            System.out.print(resultado.get(i) + ",");
+            System.out.print(resultado.get(i)+1 + ",");
         }
-        System.out.println(resultado.get(resultado.size()-1));
+        System.out.println(resultado.get(resultado.size()-1)+1);
         System.out.println(beneficio);
     }
 
@@ -94,47 +85,37 @@ public class PasteleriaDAA {
         Nodo    sol           =   new Nodo();
         Nodo    solFin        =   new Nodo();
         
-        for (int i = 0; i < pasteleros; i++) {
-            solFin.getSol().add(i+1);
-            cotaInferior+=tablaBeneficios[i][pedido[i]-1];
-        }
+        solFin = generarSol();
+        cotaInferior = cotaSuperior(solFin);
         
         PriorityQueue q = new PriorityQueue();
         q.add(sol);
         
         while (!q.isEmpty()) {
-            sol = (Nodo) q.peek();
-            q.remove();  
+            sol = (Nodo) q.remove();  
             if (esSolucion(sol)) {
-                if (sol.getBeneficio() > cotaInferior) {
+                if (cotaSuperior(sol) > cotaInferior) {
                     solFin=sol;
-                    cotaInferior=sol.getBeneficio();
+                    cotaInferior=cotaSuperior(sol);
                 }
-            } else if (cotaSuperior(sol) >= cotaInferior) {
+            } else if (cotaSuperior(sol) > cotaInferior) {
                 for (Nodo n : complecciones(sol)) {
-                    if (cotaSuperior(n)>= cotaInferior && esFactible(n)) {
+                    if (cotaSuperior(n)> cotaInferior && esFactible(n)) {
                         q.add(n);
                     }
                 }
             } 
         }
-        beneficio = solFin.getBeneficio() + sol.getPeso();
+        beneficio = cotaSuperior(solFin);
         resultado = solFin.getSol();
     }
 
     private static boolean esSolucion(Nodo sol) {
-        return sol.getNivel()==pasteleros;
+        return sol.getSol().size()==pasteleros;
     }
 
     private static int cotaSuperior(Nodo sol) {
-        int aux = sol.getBeneficio() + sol.getPeso();
-        for (int i = 0; i < pasteleros; i++) {
-            for (int j = sol.getNivel(); j < pasteleros; j++) {
-                if (!sol.getSol().contains(i)) 
-                aux+=tablaBeneficios[i][pedido[j]-1];
-            }
-        }
-        return aux;
+        
     }
 
     private static LinkedList<Nodo> complecciones(Nodo sol) {
@@ -159,5 +140,19 @@ public class PasteleriaDAA {
     private static boolean esFactible(Nodo n) {
         return true;
     }
-        
+
+    private static Nodo generarSol() {
+        Nodo nodo = new Nodo();
+        for (int i = 0; i < pasteleros; i++) {
+            nodo.getSol().add(i);
+        }
+        nodo.setNivel(pasteleros);
+        nodo.setPeso(tablaBeneficios[nodo.getSol().getLast()][pedido[pedido.length-1]-1]);
+        beneficio = 0;
+        for (int i = 0; i < pasteleros-1; i++) {
+            beneficio+=tablaBeneficios[i][pedido[i]-1];
+        }
+        nodo.setBeneficio(beneficio);
+        return nodo;
+    }    
 }
